Project: Move KubeClient to new package
