+++ pkg/kube/KubeClient.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cmd
+package kube
 
 import (
 	"context"
